Update LICENSE
I see that the PacVim project is currently indicated as both "All Rights Reserved" and "Open-Sourced via GPLv3."

Since these two differing states of release are in direct contradiction to one another, a choice needs to be made to either comply with the terms of GPLv3 or to, in the alternative, close the source code as proprietary and "reserved." If you have questions, note that I have shall more details to communicate with you before you accept or close this pull request. Thanks! zero2cx