Update Makefile
This commit increases Makefile overall coherency by simplifying the file's general flow while also eliminating some of the extraneous bits. One notable issue fixed here is that CXXFLAGS will now be used in place of CFLAGS. Second, any flags issued by the user on the command-line are now incorporated in the make/build process whereas before, these were being disregarded. Finally, there was a problem that arose when forcing the addition of "-Wall" to the build command so as to increase feedback verbosity. Doing this would, in some environments, prevent the build from linking a final executable file and the process would exit-with-error, while complaining of a missing library.

This update to Makefile was contributed by a coder (not myself) who stated in a forum discussion that they do not wish attribution for their work here. Thank you, kind coder.